feat: [fix/CameraViewReentry] 카메라 뒤로가기 시 미리보기 안보이는 오류 해결
