+++ PhotoTodo/CameraPreview.swift
@@ -10,20 +10,21 @@ import SwiftUI
 
 struct CameraPreview: UIViewRepresentable {
     
-    @ObservedObject var cameraVM: CameraViewModel
+    @State var cameraManager: CameraManager
     let frame: CGRect
     
     func makeUIView(context: Context) -> UIView {
         let view = UIViewType(frame: frame)
-        cameraVM.preview = AVCaptureVideoPreviewLayer(session: cameraVM.session)
-        cameraVM.preview.frame = frame
-        cameraVM.preview.videoGravity = .resizeAspectFill
-        view.layer.addSublayer(cameraVM.preview)
+        cameraManager.preview = AVCaptureVideoPreviewLayer(session: cameraManager.session)
+        cameraManager.preview.frame = frame
+        cameraManager.preview.videoGravity = .resizeAspectFill
+        view.layer.addSublayer(cameraManager.preview)
         return view
     }
     
     ///UIViewRepresentable를 사용하기 위한 필수요소, but 구현 시 사용 안함
     func updateUIView(_ uiView: UIViewType, context: Context) {
-        cameraVM.preview.frame = frame
+        cameraManager.preview.frame = frame
+        print("뷰 업데이트됨")
     }
 }
